refactor: geo distance functions go under Location struct
