+++ ziggurat-core-geoip/Cargo.toml
@@ -10,10 +10,10 @@ description = "GeoIP localization package for Ziggurat"
 
 [dependencies]
 async-trait = "0.1.63"
+geoutils = "0.5.1"
 ip2location = "0.3.5"
 ipgeolocate = "0.3.5"
 ipinfo = "0.5.1"
-geoutils = "0.5.1"
 rand = "0.8.5"
 serde = { version = "1", features = ["derive"] }
